Move gammapy/background/ring.py -> gammapy/cube/ring.py
