+++ gammapy/background/__init__.py
@@ -3,4 +3,3 @@
 from .background_estimate import *
 from .phase import *
 from .reflected import *
-from .ring import *
