+++ gammapy/cube/ring.py
@@ -38,7 +38,7 @@ class AdaptiveRingBackgroundEstimator:
     Example using `AdaptiveRingBackgroundEstimator`::
 
         from gammapy.maps import Map
-        from gammapy.background import AdaptiveRingBackgroundEstimator
+        from gammapy.cube import AdaptiveRingBackgroundEstimator
 
         filename = '$GAMMAPY_DATA/tests/unbundled/poisson_stats_image/input_all.fits.gz'
         images = {
